crawler/requirements.txt
fandom-py==0.2.1
Scrapy==2.6.1